Azure/tests/requirements.txt
# Core testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-timeout==2.2.0
pytest-mock==3.12.0
pytest-html==4.1.1
pytest-json-report==1.5.0

# HTTP testing
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Azure SDK
azure-cosmos==4.5.1
azure-functions==1.18.0
azure-identity==1.15.0
azure-monitor-query==1.2.0

# Data handling
feedparser==6.0.10
python-dateutil==2.8.2
pydantic==2.5.2

# AI testing (mock Anthropic)
anthropic==0.7.8

# Utilities
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0
jinja2==3.1.2

# Code quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1

